\name{aln2trj}
\alias{aln2trj}
\title{ Mapping positions from alignment to trajectory }
\description{
  Align the topology (PDB) file of a trajectory to an existing
  alignment, and then find equivalent atomic indices in the trajectory
  to those in the alignment.
}
\usage{
aln2trj(aln, pdb, inds, id="seq", file="aln2trj.fa")
}
\arguments{
  \item{aln}{ an alignment list object with \code{id} and \code{ali}
    components, similar to that generated by \code{\link{read.fasta}},
    \code{\link{read.fasta.pdb}}, and \code{\link{seqaln}}. }
  \item{pdb}{ the topology file of the trajectory in PDB format. }
  \item{inds}{ a numeric vector or a list object containing the 
        original indices for the alignment. }
  \item{id}{  keyword for searching the alignment ids to find out
     whether the sequence of the system under simulation is included 
     in the alignment.}
  \item{file}{ file name for the output alignment file; used only when
     the sequence of the simulation system is not included in the alignment. }
}
\details{
  When we compare the simulation with the PCA results of crystallographic 
  structures, we need first to find the atomic indices in the simulation 
  trajectory that correspond to the core or non-gap positions defined 
  in the PCA of crystal structures. Then we can fit the trajectory to the
  crystal structures and map the trajectory onto the same PC subspace that
  was used in the structural analysis.
}
\value{
  Returns a list of class "selection" as generated by \code{\link{atom.select}}
}
\references{
  Grant, B.J. et al. (2006) \emph{Bioinformatics} \bold{22}, 2695--2696.
}
\author{ Xin-Qiu Yao & Barry Grant }
\seealso{ \code{\link{rmsd}}, \code{\link{read.pdb}},
  \code{\link{read.fasta.pdb}}, \code{\link{read.dcd}} }
\examples{
\dontrun{
##--- Read aligned PDB coordinates (CA only)
aln  <- read.fasta(system.file("examples/kif1a.fa",package="bio3d"))
pdb.path = system.file("examples",package="bio3d")
pdbs <- read.fasta.pdb(aln, pdb.path = pdb.path, pdbext = ".ent")

##--- Read PDB coordinate for one structure (all atoms)
id <- "d2kin.1"
pdb <- read.pdb(system.file(paste("examples/", id, ".ent", sep=""), package="bio3d"))

# map the non-gap positions
pc.inds <- gap.inspect(pdbs$resno)
npc.inds <- aln2trj(aln=pdbs, pdb=pdb, inds=pc.inds$f.inds, id=id)

##--- fit the atomic coordinates to the aligned X-ray structure
xyz <- fit.xyz(pdbs$xyz[1,], pdb$xyz, atom2xyz(pc.inds$f.inds), npc.inds$xyz)

}

}
\keyword{ utilities }

