\encoding{UTF-8}
\name{centres}
\alias{centres}
\alias{centres.default}
\alias{centres.pdb}
\title{Centres-of-Geometry and Centres-of-Mass}
\description{Computes centres-of-geometry and centres-of-mass of groups of atoms.}
\usage{
centres(...)

\method{centres}{default}(x, w = NULL, factor = NULL, unsplit = FALSE, \dots)

\method{centres}{pdb}(x, w = atom2mass(x), factor = as.factor(x$atom[,"resno"]), unsplit = FALSE, \dots)

}
\arguments{
  \item{x}{an R object containing atomic coordinates.}
  \item{w}{a numeric vector containing atomic masses.}
  \item{factor}{a vector used as a factor to split the coordinates to compute multiple centres.}
  \item{unsplit}{a logical value indicating whether the coordinates of the centres have to be unsplit to repeat their coordinates for each atom used for their calculation.}
  \item{\dots}{further arguments passed to or from other methods.}
}
\details{
\code{centres} is a generic function to compute centres-of-geometry and centres-of-mass from an object containing atomic coordinates.

The default method can be used for the 'xyz' component of objects of class \sQuote{pdb}.
The coordinates of \code{x} are first splitted into groups defined by \code{factor} using the \code{\link{split}} function.
For each group, the weighted mean of the \code{x}, \code{y} and \code{z} coordinates are calculated using \code{w}.
Finally, if \code{unplit = TRUE} the coordinates of the centres are unsplitted using the \code{\link{unsplit}} function
to assign to each atom the coordinates of the centre to which they are attached.

For objects of class \sQuote{pdb} by default \code{factor} is set to \code{x$atom[,"resno"]} and atomic masses are generated by \code{atom2mass}.
Thus, the centre-of-mass of the different resdiues is computed.
Notice that coordinates can be neglected for the calculation of the centres using NA values in \code{factor}.
}
\value{
Return a 'xyz' vector containing the coordinates of the centres.
}

\author{
Julien Idé}

\seealso{
\code{\link{split}}, \code{\link{unsplit}}, \code{\link{factor}}
}
\examples{
x <- read.pdb("1bg2")
x.centres <- centres(x)
visualize(x, col = x$atom[,"resno"])
res.col <- sapply(split(x$atom[,"resno"],x$atom[,"resno"]), unique)
visualize(x.centres, type="s", col=res.col, radii=1, add=TRUE)
}
\keyword{ manip }
