\name{lbio3d}
\alias{lbio3d}
\title{ List all Functions in the bio3d Package }
\description{
  A simple shortcut for ls("package:bio3d"). 
}
\usage{
lbio3d()
}
\value{
  A character vector of function names from the bio3d package.
}
\references{
  Grant, B.J. et al. (2006) \emph{Bioinformatics} \bold{22}, 2695--2696.
}
\author{ Barry Grant }
\keyword{ utilities }
