\name{pdb2aln.ind}
\alias{pdb2aln.ind}
\title{ Mapping between PDB atomic indices and alignment positions }
\description{
  Find the best alignment between a PDB structure and an 
  existing alignment. Then, given a set of residue indices
  defined for the original alignment, return the equivalent 
  CA atom indices in the PDB coordinates.
}
\usage{
pdb2aln.ind(aln, pdb, inds = NULL, ...)
}
\arguments{
  \item{aln}{ an alignment list object with \code{id} and \code{ali}
    components, similar to that generated by \code{\link{read.fasta}},
    \code{\link{read.fasta.pdb}}, \code{\link{pdbaln}}, and 
    \code{\link{seqaln}}. }
  \item{pdb}{ the PDB object. }
  \item{inds}{ a numeric vector with a subset of the alignment position indices.  If NULL, non-gap positions is used. }
  \item{\dots}{ additional arguments for the function \code{\link{pdb2aln}}. }
}
\details{
  Call the function \code{\link{pdb2aln}} to align the PDB sequence to
  the existing alignment. Then find the equivalent CA atomic indices in 
  PDB to \code{inds}. 
}
\value{
  Returns a list containing two "select" objects:
  \item{a}{ atom and xyz indices for the alignment. }
  \item{b}{ atom and xyz indices for the PDB. }

  Note that if any element of \code{inds} has no corresponding CA atom in the 
  PDB, the output \code{a$atom} and \code{b$atom} will be shorter than 
  \code{inds}, i.e. only indices having equivalent CA atoms are returned. 
}
\references{
  Grant, B.J. et al. (2006) \emph{Bioinformatics} \bold{22}, 2695--2696.
}
\author{ Xin-Qiu Yao, Lars Skjaerven & Barry Grant }
\seealso{ \code{\link{seq2aln}}, \code{\link{seqaln.pair}},
       \code{\link{pdb2aln}} }
\examples{
\dontrun{
##--- Read aligned PDB coordinates (CA only)
aln  <- read.fasta(system.file("examples/kif1a.fa",package="bio3d"))
pdbs <- read.fasta.pdb(aln)

##--- Read PDB coordinate for a new structure (all atoms)
id <- get.pdb("2kin", URLonly=TRUE)
pdb <- read.pdb(id)

# map the non-gap positions
#npc.inds <- pdb2aln.ind(aln=pdbs, pdb=pdb, id=id)

# map the non-gap positions with known closest sequence id in the alignment
#npc.inds <- pdb2aln.ind(aln=pdbs, pdb=pdb, id=id, aln.id="1bg2")

# map core positions
core <- core.find(pdbs)
core.inds <- pdb2aln.ind(aln=pdbs, pdb=pdb, id=id, inds = core$c1A.atom)

core.inds$a
core.inds$b

##--- fit the atomic coordinates to the aligned X-ray structure
xyz <- fit.xyz(pdbs$xyz[1,], pdb$xyz, core.inds$a$xyz, core.inds$b$xyz)

}

}
\keyword{ utilities }

