\name{transducin}
\docType{data}
\alias{transducin}
\alias{pdbs}
\alias{core}
\alias{annotation}
\title{Bio3d Example Data}
\description{
  These data sets contain the results of running various bio3d
  functions on example transducin data. The main purpose of including this
  data (which may be generated by the user by following the extended
  examples documented within the various bio3d functions) is to
  speed up example execution.  It should allow users to more quickly
  appreciate the capabilities of functions that would otherwise require
  raw data input and processing before execution.
}
\usage{
data(transducin)
}
\format{
  Three objects from analysis of the \code{transducin} sequence and structure
  data:

  \enumerate{
    \item{pdbs}{is a list of class \code{"3dalign"} containing aligned PDB
      structure data.  This is the output of running
      \code{\link{read.fasta.pdb}} with the \code{aln} object. }
    \item{core}{is a list of class \code{"core"} obtained by running the
      function \code{\link{core.find}} on the \code{pdbs} object.}
    \item{annotation}{is a character matrix describing the nucleotide state and bound ligand species for each structure in \code{pdbs}.}
  }
}
\source{
  A related but more extensive dataset formed the basis of
  the work described in (Yao & Grant, 2013).
}
\references{
  Grant, B.J. et al. (2006) \emph{Bioinformatics} \bold{22}, 2695--2696.
  Yao, X.Q. et al. (2013) \emph{Biophys. J.} \bold{105}, L08--L10.
}
\keyword{datasets}
