\encoding{UTF-8}
\name{centres}
\alias{centres}
\alias{centres.default}
\alias{centres.xyz}
\alias{centres.pdb}
\title{Centres-of-Geometry and Centres-of-Mass}
\description{Computes centres-of-geometry and centres-of-mass of groups of atoms.}
\usage{
centres(...)

\method{centres}{xyz}(x, w = NULL, factor = NULL, unsplit = FALSE, \dots)

\method{centres}{pdb}(x, w = atom2mass(x), factor = as.factor(x$atom[,"resno"]), unsplit = FALSE, \dots)

}
\arguments{
  \item{x}{an R object containing atomic coordinates.}
  \item{w}{a numeric vector containing atomic masses.}
  \item{factor}{a vector used as a factor to split the coordinates to compute multiple centres.}
  \item{unsplit}{a logical value indicating whether the coordinates of the centres have to be unsplit to repeat their coordinates for each atom used for their calculation.}
  \item{\dots}{further arguments passed to or from other methods.}
}
\details{
\code{centres} is a generic function to compute centres-of-geometry and centres-of-mass from an object containing atomic coordinates.

The coordinates of \code{x} are first split into groups defined by \code{factor} using the \code{\link{split}} function.
For each group, the weighted mean of the \code{x}, \code{y} and \code{z} coordinates are calculated using \code{w} (typically an atoms mass).
Finally, if \code{unsplit = TRUE} the coordinates of the centres are ungrouped (or unsplit) using the \code{\link{unsplit}} function
to assign to each atom the coordinates of the centre to which they are attached.

For objects of class \sQuote{pdb} by default \code{factor} is set to the residue number in the PDB pbject i.e. \code{x$atom[,"resno"]} and atomic masses are generated by \code{atom2mass}.
Thus, the centre-of-mass of the different resdiues is computed.
Notice that coordinates can be neglected for the calculation of the centres using NA values in \code{factor}.
}
\value{
Return a single row 'xyz' matrix containing the coordinates of the centres.
}

\author{Julien Ide}

\seealso{
\code{\link{split}}, \code{\link{unsplit}}, \code{\link{factor}}
}
\examples{
##- Residue center of mass 
x <- atom.select(read.pdb("1bg2"), resno=10:20, value=TRUE)
x.centres <- centres(x)

##- View in interactive 3D
#view(x)
#view(x.centres, type="s", radii=0.8, col="green", add=TRUE)

}
\keyword{ manip }
